feat: use module to style to have some practice
