+++ src/components/Button/Button.module.css
@@ -1,0 +1,15 @@
+.btn {
+  padding: 10px 15px;
+  border-radius: 5px;
+  border: none;
+  margin: 5px 5px;
+}
+
+.btn-primary {
+  background-color: #188ca6;
+  color: #fff;
+}
+
+.btn:active {
+  background-color: #3f5f81;
+}
